+++ package.json
@@ -32,6 +32,7 @@
     "@nestjs/typeorm": "^9.0.1",
     "axios": "^0.27.2",
     "bcrypt": "^5.0.1",
+    "class-transformer": "^0.5.1",
     "class-validator": "^0.13.2",
     "cookie-parser": "^1.4.6",
     "dotenv": "^16.0.2",
