+++ src/app.controller.ts
@@ -3,8 +3,7 @@ import { SystemInfo } from './decorators/system-info';
 @Controller()
 export class AppController {
   @Get()
-  getHello(@SystemInfo() systemInfo, @Headers('Authorization') headers) {
-    console.log(headers);
-    return { a: 'Hell0, World!', systemInfo, headers};
+  getHello(@SystemInfo() systemInfo) {
+    return { greeting: 'Hell0, World!', systemInfo};
   }
 }
